% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ageChd.R
\docType{data}
\name{ageChd}
\alias{ageChd}
\title{Age and Coronary Heart Disease data}
\format{
A \code{data.frame} with
\eqn{100} observations (rows)
and \eqn{3} variables (columns).
}
\source{
Originally taken from \bold{H&L 2nd ed.} via their publishers site at ftp://ftp.wiley.com/public/sci_tech_med/logistic
}
\description{
Age and Coronary Heart Disease data
}
\details{
Age and presence of coronary heart disease for
100 subjects.
\cr \cr
Columns are:
\describe{
 \item{ID}{Identification code. \eqn{1} to \eqn{100}.}
 \item{age}{Age (years).}
 \item{chd}{Evidence of coronary heart disease?
            (a \code{factor}):
  \describe{
   \item{0}{no}
   \item{1}{yes}}}
}
}
\references{
\bold{H&L 2nd ed.} Page 3, Table 1.1.
}
\seealso{
\code{\link{sig}}
\code{\link{OR}}
}
\keyword{datasets}
